fix: update imports after renaming constants/languages to noteLanguages
